test fixes; formatting and other housekeeping in the wgsl
[refactoring] Separate file for the Uniforms type.
[test] Fix for the soft shadows unit test.
