+++ library/src/gpu/mod.rs
@@ -17,3 +17,4 @@ pub(crate) mod pipelines_factory;
 pub(crate) mod pipeline_code;
 pub(crate) mod adapter_features;
 mod resizable_buffer;
+pub(crate) mod uniforms;
